Enum twitch_oauth2::Scope

#[non_exhaustive]
pub enum Scope {

    AnalyticsReadExtensions,
    AnalyticsReadGames,
    BitsRead,
    ChannelEditCommercial,
    ChannelManageBroadcast,
    ChannelManageExtensions,
    ChannelManagePolls,
    ChannelManagePredictions,
    ChannelManageRedemptions,
    ChannelManageVideos,
    ChannelModerate,
    ChannelReadEditors,
    ChannelReadHypeTrain,
    ChannelReadPolls,
    ChannelReadPredictions,
    ChannelReadRedemptions,
    ChannelReadStreamKey,
    ChannelReadSubscriptions,
    ChannelSubscriptions,
    ChatEdit,
    ChatRead,
    ClipsEdit,
    ModerationRead,
    ModeratorManageAutoMod,
    UserEdit,
    UserEditBroadcast,
    UserEditFollows,
    UserManageBlockedUsers,
    UserReadBlockedUsers,
    UserReadBroadcast,
    UserReadEmail,
    UserReadFollows,
    UserReadSubscriptions,
    WhispersEdit,
    WhispersRead,
    Other(Cow<'static, str>),

}

Scopes for twitch.

https://dev.twitch.tv/docs/authentication/#scopes

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AnalyticsReadExtensions

View analytics data for the Twitch Extensions owned by the authenticated account.

analytics:read:extensions

AnalyticsReadGames

View analytics data for the games owned by the authenticated account.

analytics:read:games

BitsRead

View Bits information for a channel.

bits:read

ChannelEditCommercial

Run commercials on a channel.

channel:edit:commercial

ChannelManageBroadcast

Manage a channel’s broadcast configuration, including updating channel configuration and managing stream markers and stream tags.

channel:manage:broadcast

ChannelManageExtensions

Manage a channel’s Extension configuration, including activating Extensions.

channel:manage:extensions

ChannelManagePolls

Manage a channel’s polls.

channel:manage:polls

ChannelManagePredictions

Manage of channel’s Channel Points Predictions

channel:manage:predictions

ChannelManageRedemptions

Manage Channel Points custom rewards and their redemptions on a channel.

channel:manage:redemptions

ChannelManageVideos

Manage a channel’s videos, including deleting videos.

channel:manage:videos

ChannelModerate

Perform moderation actions in a channel. The user requesting the scope must be a moderator in the channel.

channel:moderate

ChannelReadEditors

View a list of users with the editor role for a channel.

channel:read:editors

ChannelReadHypeTrain

View Hype Train information for a channel.

channel:read:hype_train

ChannelReadPolls

View a channel’s polls.

channel:read:polls

ChannelReadPredictions

View a channel’s Channel Points Predictions.

channel:read:predictions

ChannelReadRedemptions

View Channel Points custom rewards and their redemptions on a channel.

channel:read:redemptions

ChannelReadStreamKey

View an authorized user’s stream key.

channel:read:stream_key

ChannelReadSubscriptions

View a list of all subscribers to a channel and check if a user is subscribed to a channel.

channel:read:subscriptions

ChannelSubscriptions

[DEPRECATED] Read all subscribers to your channel.

channel_subscriptions

ChatEdit

Send live stream chat and rooms messages.

chat:edit

ChatRead

View live stream chat and rooms messages.

chat:read

ClipsEdit

Manage Clips for a channel.

clips:edit

ModerationRead

View a channel’s moderation data including Moderators, Bans, Timeouts, and Automod settings.

moderation:read

ModeratorManageAutoMod

Manage messages held for review by AutoMod in channels where you are a moderator.

moderator:manage:automod

UserEdit

Manage a user object.

user:edit

UserEditBroadcast

Edit your channel’s broadcast configuration, including extension configuration. (This scope implies user:read:broadcast capability.)

user:edit:broadcast

UserEditFollows

Edit a user’s follows.

user:edit:follows

UserManageBlockedUsers

Manage the block list of a user.

user:manage:blocked_users

UserReadBlockedUsers

View the block list of a user.

user:read:blocked_users

UserReadBroadcast

View a user’s broadcasting configuration, including Extension configurations.

user:read:broadcast

UserReadEmail

Read an authorized user’s email address.

user:read:email

UserReadFollows

View the list of channels a user follows.

user:read:follows

UserReadSubscriptions

View if an authorized user is subscribed to specific channels.

user:read:subscriptions

WhispersEdit

Send whisper messages.

whispers:edit

WhispersRead

View your whisper messages.

whispers:read

Other(Cow<'static, str>)

Other scope that is not implemented.

Implementations

impl Scope

pub fn all() -> Vec<Scope>

Get a vec of all defined twitch Scopes.

Please note that this may not work for you, as some auth flows and “apis” don’t accept all scopes

pub fn description(&self) -> &'static str

Get a description for the token

pub fn parse<C>(s: C) -> Scope where
C: Into<Cow<'static, str>>,

Make a scope from a cow string

impl Scope

pub fn as_oauth_scope(&self) -> Scope

Get Scope as an oauth2 Scope