twitch_oauth2::scopes::validator
Enum Validator

#[non_exhaustive]
pub enum Validator {
    Scope(Scope),
    All(Sized<Validators>),
    Any(Sized<Validators>),
    Not(Sized<Validators>),
}

A validator is a way to check if an array of scopes matches a predicate.

Can be constructed easily with the validator! macro.

Examples

use twitch_oauth2::{validator, AppAccessToken, Scope, TwitchToken as _};

let token: AppAccessToken = token();
let validator = validator!(Scope::ChatEdit, Scope::ChatRead);
assert!(validator.matches(token.scopes()));

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Scope(Scope)

A scope

All(Sized<Validators>)

Matches true if all validators passed inside return true

Any(Sized<Validators>)

Matches true if any validator passed inside returns true

Not(Sized<Validators>)

Matches true if all validators passed inside matches false

Implementations

impl Validator

pub fn matches(&self, scopes: &[Scope]) -> bool

Checks if the given scopes match the predicate.

Examples

use twitch_oauth2::{validator, Scope};

let scopes: &[Scope] = &[Scope::ChatEdit, Scope::ChatRead];
let validator = validator!(Scope::ChatEdit, Scope::ChatRead);
assert!(validator.matches(scopes));
assert!(!validator.matches(&scopes[..1]));

pub const fn scope(scope: Scope) -> Self

Create a Validator which matches if the scope is present.

pub const fn all_multiple(ands: &'static [Validator]) -> Self

Create a Validator which matches if all validators passed inside matches true.

pub const fn any_multiple(anys: &'static [Validator]) -> Self

Create a Validator which matches if any validator passed inside matches true.

pub const fn not(not: &'static Validator) -> Self

Create a Validator which matches if all validators passed inside matches false.