Struct twitch_oauth2::UserToken

pub struct UserToken {
    pub access_token: AccessToken,
    pub login: String,
    pub user_id: String,
    pub refresh_token: Option<RefreshToken>,
    pub never_expiring: bool,
    // some fields omitted
}

An User Token from the OAuth implicit code flow or OAuth authorization code flow

Used for requests that need an authenticated user. See also AppAccessToken

See UserToken::builder for authenticating the user using the OAuth authorization code flow.

Fields
access_token: AccessToken

The access token used to authenticate requests with

login: String

Username of user associated with this token

user_id: String

User ID of the user associated with this token

refresh_token: Option<RefreshToken>

The refresh token used to extend the life of this user token

never_expiring: bool

Token will never expire

This is only true for old client IDs, like https://twitchapps.com/tmi and others

Implementations

impl UserToken

pub fn from_existing_unchecked(
access_token: impl Into<AccessToken>,
refresh_token: impl Into<Option<RefreshToken>>,
client_id: impl Into<ClientId>,
client_secret: impl Into<Option<ClientSecret>>,
login: String,
user_id: String,
scopes: Option<Vec<Scope>>,
expires_in: Option<Duration>
) -> UserToken

Assemble token without checks.

If expires_in is None, we’ll assume token.is_elapsed is always false

pub async fn from_existing<'a, C>(
http_client: &'a C,
access_token: AccessToken,
refresh_token: impl Into<Option<RefreshToken>>,
client_secret: impl Into<Option<ClientSecret>>
) -> Result<UserToken, ValidationError<<C as Client<'a>>::Error>> where
C: Client<'a>,

Assemble token and validate it. Retrieves login, client_id and scopes

If the token is already expired, this function will fail to produce a UserToken and return ValidationError::NotAuthorized

pub fn builder(
client_id: ClientId,
client_secret: ClientSecret,
redirect_url: Url
) -> UserTokenBuilder

Create a UserTokenBuilder to get a token with the OAuth Authorization Code

pub async fn mock_token<'a, C>(
http_client: &'a C,
client_id: ClientId,
client_secret: ClientSecret,
user_id: impl AsRef<str>,
scopes: Vec<Scope>
) -> Result<UserToken, UserTokenExchangeError<<C as Client<'a>>::Error>> where
C: Client<'a>,

This is supported on crate feature mock_api only.

Generate a user token from mock-api

Examples

let token = twitch_oauth2::UserToken::mock_token(
    &reqwest::Client::builder()
        .redirect(reqwest::redirect::Policy::none())
        .build()?,
    "mockclientid".into(),
    "mockclientsecret".into(),
    "user_id",
    vec![],
    ).await?;

pub fn set_secret(&mut self, secret: Option<ClientSecret>)

Set the client secret