Struct twitch_oauth2::tokens::UserTokenBuilder

pub struct UserTokenBuilder { /* private fields */ }

Builder for OAuth authorization code flow

See ImplicitUserTokenBuilder for the OAuth implicit code flow (does not require Client Secret)

Implementations

impl UserTokenBuilder

pub fn new(
client_id: ClientId,
client_secret: ClientSecret,
redirect_url: Url
) -> UserTokenBuilder

Create a UserTokenBuilder

Notes

The url crate converts empty paths into “/” (such as https://example.com into https://example.com/), which means that you’ll need to add https://example.com/ to your redirect URIs (note the “trailing” slash) if you want to use an empty path.

To avoid this, use a path such as https://example.com/twitch/register or similar instead, where the url crate would not add a trailing /.

pub fn set_scopes(self, scopes: Vec<Scope>) -> Self

Add scopes to the request

pub fn add_scope(&mut self, scope: Scope)

Add a single scope to request

pub fn force_verify(self, b: bool) -> Self

Enable or disable function to make the user able to switch accounts if needed.

pub fn generate_url(&mut self) -> (Url, CsrfToken)

Generate the URL to request a code.

Step 1. in the guide

pub async fn get_user_token<'a, C>(
self,
http_client: &'a C,
state: &str,
code: &str
) -> Result<UserToken, UserTokenExchangeError<<C as Client<'a>>::Error>> where
C: Client<'a>,

Generate the code with the help of the authorization code

Step 3. and 4. in the guide

On failure to authenticate due to wrong redirect url or other errors, twitch redirects the user to <redirect_url or first defined url in dev console>?error=<error type>&error_description=<description of error>