Enum twitch_oauth2::Scope

#[non_exhaustive]
pub enum Scope {

    AnalyticsReadExtensions,
    AnalyticsReadGames,
    BitsRead,
    ChannelEditCommercial,
    ChannelManageBroadcast,
    ChannelManageExtensions,
    ChannelManageModerators,
    ChannelManagePolls,
    ChannelManagePredictions,
    ChannelManageRaids,
    ChannelManageRedemptions,
    ChannelManageSchedule,
    ChannelManageVideos,
    ChannelManageVips,
    ChannelModerate,
    ChannelReadCharity,
    ChannelReadEditors,
    ChannelReadGoals,
    ChannelReadHypeTrain,
    ChannelReadPolls,
    ChannelReadPredictions,
    ChannelReadRedemptions,
    ChannelReadStreamKey,
    ChannelReadSubscriptions,
    ChannelReadVips,
    ChannelSubscriptions,
    ChatEdit,
    ChatRead,
    ClipsEdit,
    ModerationRead,
    ModeratorManageAnnouncements,
    ModeratorManageAutoMod,
    ModeratorManageAutomodSettings,
    ModeratorManageBannedUsers,
    ModeratorManageBlockedTerms,
    ModeratorManageChatMessages,
    ModeratorManageChatSettings,
    ModeratorReadAutomodSettings,
    ModeratorReadBlockedTerms,
    ModeratorReadChatSettings,
    UserEdit,
    UserEditBroadcast,
    UserEditFollows,
    UserManageBlockedUsers,
    UserManageChatColor,
    UserManageWhispers,
    UserReadBlockedUsers,
    UserReadBroadcast,
    UserReadEmail,
    UserReadFollows,
    UserReadSubscriptions,
    WhispersEdit,
    WhispersRead,
    Other(Cow<'static, str>),

}

Scopes for twitch.

https://dev.twitch.tv/docs/authentication/#scopes

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AnalyticsReadExtensions

View analytics data for the Twitch Extensions owned by the authenticated account.

analytics:read:extensions

AnalyticsReadGames

View analytics data for the games owned by the authenticated account.

analytics:read:games

BitsRead

View Bits information for a channel.

bits:read

ChannelEditCommercial

Run commercials on a channel.

channel:edit:commercial

ChannelManageBroadcast

Manage a channel’s broadcast configuration, including updating channel configuration and managing stream markers and stream tags.

channel:manage:broadcast

ChannelManageExtensions

Manage a channel’s Extension configuration, including activating Extensions.

channel:manage:extensions

ChannelManageModerators

Add or remove the moderator role from users in your channel.

channel:manage:moderators

ChannelManagePolls

Manage a channel’s polls.

channel:manage:polls

ChannelManagePredictions

Manage of channel’s Channel Points Predictions

channel:manage:predictions

ChannelManageRaids

Manage a channel raiding another channel.

channel:manage:raids

ChannelManageRedemptions

Manage Channel Points custom rewards and their redemptions on a channel.

channel:manage:redemptions

ChannelManageSchedule

Manage a channel’s stream schedule.

channel:manage:schedule

ChannelManageVideos

Manage a channel’s videos, including deleting videos.

channel:manage:videos

ChannelManageVips

Add or remove the VIP role from users in your channel.

channel:manage:vips

ChannelModerate

Perform moderation actions in a channel. The user requesting the scope must be a moderator in the channel.

channel:moderate

ChannelReadCharity

Read charity campaign details and user donations on your channel.

channel:read:charity

ChannelReadEditors

View a list of users with the editor role for a channel.

channel:read:editors

ChannelReadGoals

View Creator Goals for a channel.

channel:read:goals

ChannelReadHypeTrain

View Hype Train information for a channel.

channel:read:hype_train

ChannelReadPolls

View a channel’s polls.

channel:read:polls

ChannelReadPredictions

View a channel’s Channel Points Predictions.

channel:read:predictions

ChannelReadRedemptions

View Channel Points custom rewards and their redemptions on a channel.

channel:read:redemptions

ChannelReadStreamKey

View an authorized user’s stream key.

channel:read:stream_key

ChannelReadSubscriptions

View a list of all subscribers to a channel and check if a user is subscribed to a channel.

channel:read:subscriptions

ChannelReadVips

Read the list of VIPs in your channel.

channel:read:vips

ChannelSubscriptions

[DEPRECATED] Read all subscribers to your channel.

channel_subscriptions

ChatEdit

Send live stream chat and rooms messages.

chat:edit

ChatRead

View live stream chat and rooms messages.

chat:read

ClipsEdit

Manage Clips for a channel.

clips:edit

ModerationRead

View a channel’s moderation data including Moderators, Bans, Timeouts, and Automod settings.

moderation:read

ModeratorManageAnnouncements

Send announcements in channels where you have the moderator role.

moderator:manage:announcements

ModeratorManageAutoMod

Manage messages held for review by AutoMod in channels where you are a moderator.

moderator:manage:automod

ModeratorManageAutomodSettings

Manage a broadcaster’s AutoMod settings

moderator:manage:automod_settings

ModeratorManageBannedUsers

Ban and unban users.

moderator:manage:banned_users

ModeratorManageBlockedTerms

Manage a broadcaster’s list of blocked terms.

moderator:manage:blocked_terms

ModeratorManageChatMessages

Delete chat messages in channels where you have the moderator role

moderator:manage:chat_messages

ModeratorManageChatSettings

View a broadcaster’s chat room settings.

moderator:manage:chat_settings

ModeratorReadAutomodSettings

View a broadcaster’s AutoMod settings.

moderator:read:automod_settings

ModeratorReadBlockedTerms

View a broadcaster’s list of blocked terms.

moderator:read:blocked_terms

ModeratorReadChatSettings

View a broadcaster’s chat room settings.

moderator:read:chat_settings

UserEdit

Manage a user object.

user:edit

UserEditBroadcast

Edit your channel’s broadcast configuration, including extension configuration. (This scope implies user:read:broadcast capability.)

user:edit:broadcast

UserEditFollows

[DEPRECATED] Was previously used for “Create User Follows” and “Delete User Follows.

user:edit:follows

UserManageBlockedUsers

Manage the block list of a user.

user:manage:blocked_users

UserManageChatColor

Update the color used for the user’s name in chat.Update User Chat Color

user:manage:chat_color

UserManageWhispers

Read whispers that you send and receive, and send whispers on your behalf.

user:manage:whispers

UserReadBlockedUsers

View the block list of a user.

user:read:blocked_users

UserReadBroadcast

View a user’s broadcasting configuration, including Extension configurations.

user:read:broadcast

UserReadEmail

View a user’s email address.

user:read:email

UserReadFollows

View the list of channels a user follows.

user:read:follows

UserReadSubscriptions

View if an authorized user is subscribed to specific channels.

user:read:subscriptions

WhispersEdit

Send whisper messages.

whispers:edit

WhispersRead

View your whisper messages.

whispers:read

Other(Cow<'static, str>)

Other scope that is not implemented.

Implementations

impl Scope

pub fn all() -> Vec<Scope>

Get a vec of all defined twitch Scopes.

Please note that this may not work for you, as some auth flows and “apis” don’t accept all scopes

pub fn description(&self) -> &'static str

Get a description for the token

pub fn parse<C>(s: C) -> Scopewhere
C: Into<Cow<'static, str>>,

Make a scope from a cow string

pub fn as_str(&self) -> &str

Get the scope as a borrowed string.