Struct twitch_oauth2::AppAccessToken

pub struct AppAccessToken {
    pub access_token: AccessToken,
    pub refresh_token: Option<RefreshToken>,
    /* private fields */
}

An App Access Token from the OAuth client credentials flow

Used for server-to-server requests. Use UserToken for requests that need to be in the context of an authenticated user.

In some contexts (i.e EventSub) an App Access Token can be used in the context of users that have authenticated the specific Client ID

Fields
access_token: AccessToken

The access token used to authenticate requests with

refresh_token: Option<RefreshToken>

The refresh token used to extend the life of this user token

Implementations

impl AppAccessToken

pub fn from_existing_unchecked(
access_token: AccessToken,
refresh_token: impl Into<Option<RefreshToken>>,
client_id: impl Into<ClientId>,
client_secret: impl Into<ClientSecret>,
scopes: Option<Vec<Scope>>,
expires_in: Option<Duration>
) -> AppAccessToken

Assemble token without checks.

If expires_in is None, we’ll assume token.is_elapsed() == true

pub async fn from_existing<'a, RE, C>(
http_client: &'a C,
access_token: AccessToken,
refresh_token: impl Into<Option<RefreshToken>>,
client_secret: ClientSecret
) -> Result<AppAccessToken, ValidationError<<C as Client<'a>>::Error>> where
C: Client<'a>,

Assemble token and validate it. Retrieves client_id and scopes.

pub async fn get_app_access_token<'a, C>(
http_client: &'a C,
client_id: ClientId,
client_secret: ClientSecret,
scopes: Vec<Scope>
) -> Result<AppAccessToken, AppAccessTokenError<<C as Client<'a>>::Error>> where
C: Client<'a>,

Generate app access token via OAuth client credentials flow