twitch_oauth2::tokens

Struct UserToken

pub struct UserToken {
    pub access_token: AccessToken,
    pub login: UserName,
    pub user_id: UserId,
    pub refresh_token: Option<RefreshToken>,
    pub never_expiring: bool,
    /* private fields */
}

An User Token from the OAuth implicit code flow or OAuth authorization code flow

Used for requests that need an authenticated user. See also AppAccessToken

See UserToken::builder for authenticating the user using the OAuth authorization code flow.

Fields
access_token: AccessToken

The access token used to authenticate requests with

login: UserName

Username of user associated with this token

user_id: UserId

User ID of the user associated with this token

refresh_token: Option<RefreshToken>

The refresh token used to extend the life of this user token

never_expiring: bool

Token will never expire

This is only true for old client IDs, like https://twitchapps.com/tmi and others

Implementations

impl UserToken

pub fn new( access_token: AccessToken, refresh_token: Option<RefreshToken>, validated: ValidatedToken, client_secret: impl Into<Option<ClientSecret>>, ) -> Result<UserToken, ValidationError<Infallible>>

Create a new token

See UserToken::from_token and UserToken::from_existing for more ways to create a UserToken

pub async fn from_token<C>( http_client: &C, access_token: AccessToken, ) -> Result<UserToken, ValidationError<<C as Client>::Error>>
where C: Client,

Available on crate feature client only.

Create a UserToken from an existing active user token. Retrieves login, client_id and scopes

If the token is already expired, this function will fail to produce a UserToken and return ValidationError::NotAuthorized

Examples

use twitch_oauth2::{AccessToken, UserToken};
// Make sure you enable the feature "reqwest" for twitch_oauth2 if you want to use reqwest
let client = reqwest::Client::builder()
    .redirect(reqwest::redirect::Policy::none())
    .build()?;
let token = UserToken::from_token(&client, AccessToken::from("my_access_token")).await?;

pub async fn from_existing<C>( http_client: &C, access_token: AccessToken, refresh_token: impl Into<Option<RefreshToken>>, client_secret: impl Into<Option<ClientSecret>>, ) -> Result<UserToken, ValidationError<<C as Client>::Error>>
where C: Client,

Available on crate feature client only.

Create a UserToken from an existing active user token. Retrieves login, client_id and scopes

If the token is already expired, this function will fail to produce a UserToken and return ValidationError::NotAuthorized

Examples

use twitch_oauth2::{AccessToken, ClientSecret, RefreshToken, UserToken};
// Make sure you enable the feature "reqwest" for twitch_oauth2 if you want to use reqwest
let client = reqwest::Client::builder()
    .redirect(reqwest::redirect::Policy::none())
    .build()?;
let token = UserToken::from_existing(
    &client,
    AccessToken::from("my_access_token"),
    RefreshToken::from("my_refresh_token"),
    ClientSecret::from("my_client_secret"),
)
.await?;

pub fn from_existing_unchecked( access_token: impl Into<AccessToken>, refresh_token: impl Into<Option<RefreshToken>>, client_id: impl Into<ClientId>, client_secret: impl Into<Option<ClientSecret>>, login: UserName, user_id: UserId, scopes: Option<Vec<Scope>>, expires_in: Option<Duration>, ) -> UserToken

Assemble token without checks.

Notes

If expires_in is None, we’ll assume token.is_elapsed is always false

pub fn from_response( response: TwitchTokenResponse, validated: ValidatedToken, client_secret: impl Into<Option<ClientSecret>>, ) -> Result<UserToken, ValidationError<Infallible>>

Assemble token from twitch responses.

pub fn builder( client_id: ClientId, client_secret: ClientSecret, redirect_url: Url, ) -> UserTokenBuilder

Create a UserTokenBuilder to get a token with the OAuth Authorization Code

pub async fn mock_token<C>( http_client: &C, client_id: ClientId, client_secret: ClientSecret, user_id: impl AsRef<str>, scopes: Vec<Scope>, ) -> Result<UserToken, UserTokenExchangeError<<C as Client>::Error>>
where C: Client,

Available on crate features mock_api and client only.

Generate a user token from mock-api

Examples

let token = twitch_oauth2::UserToken::mock_token(
    &reqwest::Client::builder()
        .redirect(reqwest::redirect::Policy::none())
        .build()?,
    "mockclientid".into(),
    "mockclientsecret".into(),
    "user_id",
    vec![],
)
.await?;

pub fn set_secret(&mut self, secret: Option<ClientSecret>)

Set the client secret