Struct twitch_oauth2::tokens::UserTokenBuilder

pub struct UserTokenBuilder { /* private fields */ }

Builder for OAuth authorization code flow

See ImplicitUserTokenBuilder for the OAuth implicit code flow (does not require Client Secret)

Implementations

impl UserTokenBuilder

pub fn new( client_id: impl Into<ClientId>, client_secret: impl Into<ClientSecret>, redirect_url: Url ) -> UserTokenBuilder

Create a UserTokenBuilder

Notes

The url crate converts empty paths into “/” (such as https://example.com into https://example.com/), which means that you’ll need to add https://example.com/ to your redirect URIs (note the “trailing” slash) if you want to use an empty path.

To avoid this, use a path such as https://example.com/twitch/register or similar instead, where the url crate would not add a trailing /.

pub fn set_scopes(self, scopes: Vec<Scope>) -> Self

Add scopes to the request

pub fn add_scope(&mut self, scope: Scope)

Add a single scope to request

pub fn force_verify(self, b: bool) -> Self

Enable or disable function to make the user able to switch accounts if needed.

pub fn generate_url(&mut self) -> (Url, CsrfToken)

Generate the URL to request a code.

Step 1. in the guide

pub fn csrf_is_valid(&self, csrf: &str) -> bool

Check if the CSRF is valid

pub fn get_user_token_request(&self, code: &str) -> Request<Vec<u8>>

Get the request for getting a TwitchTokenResponse, to be used in UserToken::from_response.

Examples

use twitch_oauth2::{tokens::UserTokenBuilder, id::TwitchTokenResponse};
use url::Url;
let callback_url = Url::parse("http://localhost/twitch/register")?;
let mut builder = UserTokenBuilder::new("myclientid", "myclientsecret", callback_url);
let (url, _csrf_code) = builder.generate_url();

// Direct the user to this url.
// Later when your server gets a response on `callback_url` with `?code=xxxxxxx&state=xxxxxxx&scope=aa%3Aaa+bb%3Abb`

// validate the state
if !builder.csrf_is_valid(state_in_query) {
    panic!("state mismatched")
}
// and then get your token
let request = builder.get_user_token_request(code_in_query);

// use your favorite http client

let response: http::Response<Vec<u8>> = client_req(request);
let twitch_response = TwitchTokenResponse::from_response(&response)?;

// you now have a access token, do what you want with it.
// You're recommended to convert it into a `UserToken` via `UserToken::from_response`

// You can validate the access_token like this
let validated_req = twitch_response.access_token.validate_token_request();

pub async fn get_user_token<'a, C>( self, http_client: &'a C, state: &str, code: &str ) -> Result<UserToken, UserTokenExchangeError<<C as Client>::Error>>where C: Client,

Available on crate feature client only.

Generate the code with the help of the authorization code

Step 3. and 4. in the guide

On failure to authenticate due to wrong redirect url or other errors, twitch redirects the user to <redirect_url or first defined url in dev console>?error=<error type>&error_description=<description of error>